version update

+++ lib/core_extend/big_decimal.rb
@@ -1,4 +1,3 @@
-class Float
+class BigDecimal
   include CoreExtend::FloatModule
-
 end
